moved LecHomePage to components
